build: disable sanitizers for shared libraries
Shared libraries built with sanitizers are a bit inconvenient to use
because they require that any application linking to them is run with
libasan preloaded using LD_PRELOAD. This limitation makes the
sanitizer support less useful because applications will refuse to
start unless there is a special environment variable set.

It seems more useful to have sanitizers enabled only for executables
and not for libraries.